requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
colorama==0.4.6
Django==4.1.4
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
